requirements.txt
scikit-learn
sentence-transformers
selenium
requests
python-dotenv